mysite/requirements.txt
alabaster==0.7.9
Babel==2.3.4
beautifulsoup4==4.5.1
Django==1.10
django-appconf==1.0.2
django-compress==1.0.1
django-compressor==2.1
django-flat-theme==1.1.3
docutils==0.12
Flat==0.1
gunicorn==19.6.0
imagesize==0.7.1
Jinja2==2.8
lxml==3.6.4
MarkupSafe==0.23
MySQL-python==1.2.5
Pygments==2.1.3
python-dateutil==2.5.3
pytz==2016.6.1
rcssmin==1.0.6
rjsmin==1.0.12
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.4.6
